requirements.txt
reportlab>=3.6.0
qrcode>=7.4.2
pillow>=9.0.0